AC 19102018 1000 CostHistoryUI is fixed

+++ p3s/src/main/java/com/bcs/p3s/service/Form1200ServiceImpl.java
@@ -232,7 +232,7 @@ public class Form1200ServiceImpl extends ServiceAuthorisationTools implements Fo
 		long range = clicksEnd - clicksStart;
 		long lngPercent = (elapsed * 100L)	 / range; 
 		
-		return (int) lngPercent;
+		return ((Long) lngPercent).intValue();
 	}
 
 	/**
